python/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
python-dotenv==1.0.1
pydantic==2.10.4
anthropic==0.44.0
aiosqlite==0.21.0
python-multipart==0.0.20
